% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.cast_NA.r
\name{ds.cast_NA}
\alias{ds.cast_NA}
\title{Cast "NA" string into NA}
\usage{
ds.cast_NA(df, modified_obj = NULL, datasources = NULL)
}
\arguments{
\item{df}{A character string specifying the name of the server-side data frame to be converted.}

\item{modified_obj}{A character string specifying the name of the new server-side
object to store the modified data frame. If NULL, a name is generated automatically
by appending \code{"_modified"} to \code{df}.}

\item{datasources}{A list of \code{\link[DSI]{DSConnection-class}} objects obtained
after login. If the \code{datasources} argument is not specified the default set of
connections will be used: see \code{\link[DSI]{datashield.connections_default}}.}
}
\value{
Returns the name of the newly created server-side object
containing the modified data frame.
}
\description{
Executes \code{cast_NADS} server-side function to convert "NA" strings into NA values across a dataset,
and reports whether any replacements occurred.
}
