% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.check_variables.r
\name{ds.check_variables}
\alias{ds.check_variables}
\title{Check variables on a server-side data frame}
\usage{
ds.check_variables(df, variables, datasources = NULL)
}
\arguments{
\item{df}{A character string specifying the name of the server-side data frame
to check.}

\item{variables}{A character vector specifying the expected variable names.}

\item{datasources}{A list of \code{\link[DSI]{DSConnection-class}} objects obtained
after login. If the \code{datasources} argument is not specified, the default set of
connections will be used: see \code{\link[DSI]{datashield.connections_default}}.}
}
\value{
Invisible list of results returned by the server-side function. Each element
corresponds to a server and contains two components: \code{missing} (variables
present in \code{variables} but missing in the data frame) and \code{extra}
(variables present in the data frame but not in \code{variables}).
}
\description{
Verify that a server-side data frame contains exactly the expected
variables. Uses the server-side function \code{check_variablesDS} to identify
missing or extra variables. Messages are printed for each server indicating
discrepancies.

Server function called: \code{check_variablesDS}
}
