% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.global_imputation.R
\name{ds.global_imputation}
\alias{ds.global_imputation}
\title{Global imputation of a server-side object using MICE}
\usage{
ds.global_imputation(
  df,
  id_col = NULL,
  m = 5,
  maxit = 30,
  imputed_newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{df}{A character string specifying the name of the server-side data frame
on which the global imputation is performed.}

\item{id_col}{A character string indicating the name of the identifier column that
is excluded from imputation.}

\item{m}{A numeric value representing the number of imputations to generate.
Default is 5.}

\item{maxit}{A numeric value representing the maximum number of MICE iterations per imputation chain.
Default is 30.}

\item{imputed_newobj}{A character string specifying the name of the new
imputed object to be created on each server. If NULL, a name is generated
automatically by appending \code{"_imputed"} to \code{x}.}

\item{datasources}{A list of DataSHIELD connections. If NULL, the active
connections returned by \code{datashield.connections_find()} are used.}
}
\value{
The name of the newly created server-side object containing the
imputed data set.
}
\description{
Performs imputation on data from a single source on the server-side.

Server function called: \code{global_imputationDS}
}
