% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.check_missing_data.R
\name{ds.check_missing_data}
\alias{ds.check_missing_data}
\title{Filter columns of a server-side object based on amissingness threshold}
\usage{
ds.check_missing_data(
  df,
  threshold,
  type = c("combined", "split"),
  preproc_newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{df}{A character string providing the name of the input data frame object
stored on each server.}

\item{threshold}{A numeric value between 0 and 100 representing the maximum
acceptable percentage of missing values per column.}

\item{type}{A character string specifying the filtering strategy. Must be one
of \code{"combined"} or \code{"split"}.}

\item{preproc_newobj}{A character string specifying the name of the new
pre-processed object to be created on each server. If \code{NULL}, a name is
generated automatically by appending \code{"_preprocessed"} to \code{df}.}

\item{datasources}{a list of \code{\link[DSI]{DSConnection-class}}
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link[DSI]{datashield.connections_default}}.}
}
\value{
A character string giving the name of the newly created server-side
object containing the pre-processed data set. If no columns are removed, the
original data are left unchanged and the name of the intended output object
is still returned invisibly.
}
\description{
Filters the columns of a server-side data set by removing variables whose
proportion of missing values exceeds a user-defined threshold.
The function can operate in two modes:
\itemize{
\item \code{"combined"}: computes a weighted global missingness across all
servers and removes the same set of variables from every server.
\item \code{"split"}: applies the threshold independently on each server and
removes server-specific sets of variables.
}

If no variables exceed the threshold, no filtering is applied and a message
is returned indicating that no columns need to be removed.

Server function called: \code{preprocess_dataDS}, \code{remove_columnDS}
}
